pyproject.toml
[project]
name = "ada"
version = "0.1.0"
description = "ADA: Analyze CRM contacts and produce actionable lead insights."
requires-python = ">=3.11"
dependencies = [
  "httpx",
  "pydantic",
  "tenacity",
  "rich",
  "pandas",
  "tabulate",
  "pyyaml"
]

[tool.setuptools]
py-modules = []

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["." ]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "UP", "B"]
ignore = []
fix = true

[tool.ruff.lint.isort]
known-first-party = ["ada"]

[tool.black]
line-length = 100
target-version = ["py311"]
